chore: refactor cluster config provider
Extract ClusterConfig's methods into own file to make it slightly more manageable.
